+++ SandboxiePlus/SandMan/SbiePlusAPI.h
@@ -146,6 +146,7 @@ public:
 	{
 		eHardenedPlus,
 		eHardened,
+		eIsoationMax,
 		eDefaultPlus,
 		eDefault,
 		eAppBoxPlus,